+++ usr.sbin/bhyve/virtio.h
@@ -23,7 +23,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $FreeBSD: head/usr.sbin/bhyve/virtio.h 264055 2014-04-02 20:18:17Z grehan $
+ * $FreeBSD: head/usr.sbin/bhyve/virtio.h 268276 2014-07-05 02:38:53Z grehan $
  */
 
 #ifndef	_VIRTIO_H_
@@ -352,7 +352,7 @@ struct virtio_consts {
 					/* called to read config regs */
 	int	(*vc_cfgwrite)(void *, int, int, uint32_t);
 					/* called to write config regs */
-	uint32_t vc_hv_caps;		/* hypervisor-provided capabilities */
+	uint64_t vc_hv_caps;		/* hypervisor-provided capabilities */
 };
 
 /*
